requirements.txt
asgiref==3.5.2
Django==3.2.15
django-jalali-date==1.0.1
jdatetime==4.1.0
psycopg2==2.9.3
pytz==2022.2.1
sqlparse==0.4.3
